azure-openai/requirements.txt
azure-functions==1.23.0
slack_sdk==3.35.0
openai==0.28.1
requests==2.32.3
